log is imported

+++ src/main/python/widgets/dragDropListView.py
@@ -1,3 +1,4 @@
+import log
 from PySide2 import QtWidgets, QtCore
 from widgets.toast import Toast
 
@@ -27,11 +28,12 @@ class DragDropListView(QtWidgets.QListView):
 			urls = event.mimeData().text().split('\n')
 			model = self.model()
 			if model is not None:
-				notContainsUrl = list(filter(lambda url: model.contains(lambda songs: songs.getUrl() == url) is False, urls))
+				notContainsUrl = list(
+					filter(lambda url: model.contains(lambda songs: songs.getUrl() == url) is False, urls))
 				if notContainsUrl:
 					self.dropped.emit(notContainsUrl)
 		except Exception as e:
-			print(e)
+			log.error(f'Drop event has error in drag drop list view. Exception is {e}')
 
 
 	def dragEnterEvent(self, event):
@@ -49,13 +51,16 @@ class DragDropListView(QtWidgets.QListView):
 		delete = contextMenu.addAction('Remove')
 
 		action = contextMenu.exec_(globalPos)
-
-		if action == delete:
-			self.__deleteSelectedItems(self.selectedIndexes())
+		if self.selectedIndexes():
+			if action == delete:
+				self.__deleteSelectedItems(self.selectedIndexes())
 
 
 	def __deleteSelectedItems(self, indexes):
-		model = self.model()
-		result = model.deleteRows(indexes)
-		if result is False:
-			Toast.error('Delete Song Item Error', 'Song is not deleted successfully')
+		res = QtWidgets.QMessageBox.question(self, 'Delete Songs', 'Are you sure that',
+											 QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
+		if res == QtWidgets.QMessageBox.Yes:
+			model = self.model()
+			result = model.deleteRows(indexes)
+			if result is False:
+				Toast.error('Delete Song Item Error', 'Song is not deleted successfully')
